fix serverless open files by disabling ssr on dog page
